requirements.txt
aiogram==3.22.0
python_decouple==3.8.0